autoscoper now loads a trial without segfaulting: * fixed order of OpenCL/OpenGL context initialization on Linux * fixed bug in order of init() and render() calls in View * fixed double free in Video destructor by adding a deep copy constructor   for TiffImage * fixed compile errors in RayCaster.cl * bug to fix: render output from RayCaster is not correct

+++ libautoscoper/RayCaster.cpp
@@ -143,7 +143,7 @@ RayCaster::setViewport(float x, float y, float width, float height)
 }
 
 void
-RayCaster::render(const Buffer* buffer, size_t width, size_t height)
+RayCaster::render(const Buffer* buffer, unsigned width, unsigned height)
 {
     if (!volumeDescription_) {
         cerr << "RayCaster: WARNING: No volume loaded. " << endl;
